+++ app.config.js
@@ -1,0 +1,16 @@
+'use strict';
+
+module.exports = {
+   'D:/Nodejs/Project/ioa/@ioa/http': {
+      "enable": true,
+   },
+   "./component/admin": {
+      "enable": true
+   },
+   "./component/other": {
+      "enable": true
+   },
+   "./component/user": {
+      "enable": true
+   },
+}