+++ app/.loader.js
@@ -1,0 +1,10 @@
+'use strict';
+
+const loader = require('D:/Nodejs/Project/ioa/@ioa/http/loader');
+
+module.exports = {
+   ...loader,
+   other: {
+      level: 16
+   }
+}