+++ app/middleware/cors.js
@@ -3,7 +3,7 @@
 /**
  * package形态的中间件
  */
- 
+
 const cors = require('@koa/cors')
 
-module.exports = cors({ origin: '*' })
+module.exports = cors({ origin: '*' });