+++ apps/admin/router.js
@@ -1,9 +1,0 @@
-'use strict';
-
-const { router, middleware } = require('@app')
-
-const { cors } = middleware
-
-router.get('/admin', cors, 'index.index')
-
-router.get('/admin/:id', cors, 'index.details')