+++ apps/main/.loader.js
@@ -1,7 +1,0 @@
-'use strict';
-
-module.exports = {
-   other: {
-      level: 16
-   }
-}