+++ apps/user/.import.js
@@ -1,7 +1,0 @@
-'use strict';
-
-module.exports = {
-   middleware: {
-      cors: "main.middleware"
-   }
-}